refactor(server): move `utils.ts` to `utils/process.ts`
